Incomplete - task 31: Address Correction Engine 'debug mode' does not rollback changes if 'immediately commit auto-corrected addresses' is checked","Type-Defect, Priority-High, OpSys-All, Milestone-Release0.9.6, DBMS-SQLServer, Component-Logic, Component-Persistence http://code.google.com/p/power-matchmaker/issues/detail?id=31
The transformation steps 'rollback' method never got called in debug mode (but would get called in the event of an exception). So debug mode never rolled back its changes in the Address Correction mode. It's also possible (looking at the code) that this also affects Cleanse projects, but I'll have to confirm.