Fixed the DQguru's 'Show Command' button so that it displays a more proper command message.
This will typically work for Windows and Linux so far, but because OS X is packaged in an App Bundle, I'm not sure if it is correct for OS X yet.

One refactoring I did do was remove the command assembly code from the ShowCommandAction. I don't see why that logic should be in the a Swing Action, and I think it makes way more sense for that to be in a core class.