NEW - bug 1778: Previewing a previously created and saved munge process of a cleansing project causes the results of the munge process for the previewed rows to be written into the source table http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=1778
The root cause was the MungePreviewer was being created before the steps were defined to be in preview mode and that the MungePreviewer was running the process in its constructor for a first set of data. This caused the previewer to execute the cleanse step without rolling it back and changing the database. Now the preview data's initial retrieval is done when its refresh is enabled as that will be the first time it will be displayed reducing effort in the constructor.