Incomplete - task 29: Using the Preview feature in the Transformation editor while running a Transformation throws an NPE","Type-Defect, Priority-High, Milestone-Release1.0, OpSys-All, Component-UI http://code.google.com/p/power-matchmaker/issues/detail?id=29
Added a new event system that will allow client code to be notified when a MatchMakerEngine is starting and stopping. This allows GUI components to enable/disable features appropriately when the engines are running/stopping. In particular for this bug, I've disabled the Preview checkbox while the engine is running. I've also used this system to disable the Run Engine Button in the EngineSettingsPanel when an engine is running, and renable it when it's stopped.