This fixes an issue where when starting up DQguru on a user account that's never used DQguru or Power Architect (more on why that matters later) resulted in the program popping up a dialogue asking for a pl.ini for DQguru, and then another one after asking for one for Architect.
The reason this was happening was because when the DQguru tries to create the default repository, it uses the Architect core to open an Architect data model of the default repository, and then forward engineer it into a local HSQLDB instance. But the Architect Core classes, before forward engineering, would look for the pl.ini location in the Architect preferences, and if the Architect had never been run on that machine, it would ask the user for it. The solution committed is to pass the DQguru's preferred pl.ini location to the Architect core code and use that instead, or if null, then use the default pl.ini location.