My previous commit to upgrade the MySQL connector JDBC driver from 5.1.5 to 5.1.12 has undesired behaviour. It causes users' MySQL driver in any SQL Power application that uses it to go "missing". The user would have to change their user preferences to relocate the MySQL driver after the upgrade.
After a discussion with Jonathan, we agreed the quickest (but not the best) fix for this (and for any future driver upgrades) is to remove the version number from the filename. Then for any future driver upgrades, we would simply replace the old driver with the new one, keeping the same filename. However, to maintain compatibility with older applications which expect filenames with versions in them, we must keep the legacy drivers in place. Thus, we would have at most 2 versions of the driver: the legacy driver (with version in filename) and the most recent driver (without version in filename).