Corrected some annotations relating to FolderParents for persistence as well as getting the FolderParentTests to pass.
To get the FolderParentTests to work the session had to be removed from the constructor of the FolderParent. There is no known reason why the folder parent had a session since the MatchMakerObject already has a session and should walk up the tree to the MMRootNode to get its session.

For simplicity the session has been removed from all of the duplicate methods. It seems less confusing if one root object holds a pointer to the correct session instead of every object holding a pointer to a session that could point to different sessions.