A large refactoring to convert a copy of MatchMaker into a smaller application that can create xml files based on projects in a repository.
Most of the UI has been stripped out of this project because it only needs the login dialog to log into a repository to export it. Most of the object model still exists as Hibernate needs it to create all of the objects before exporting them to XML. 

The approach to make a stand-alone application to export the xml is for several reasons. The idea to include this code in the real MatchMaker meant some properties could not be made appropriately final. The idea to include this code as a jar and call a utility method from the real MatchMaker failed due to class-loader hell as both this project and the real MatchMaker had the same core object names but Hibernate appears to want the names to stay as is but the objects are loaded from the real MatchMaker which no longer have the necessary setters for all of the Hibernate properties.

The UI changes to this project are not yet complete.