Merged in the spobject-conversion branch.
The merged branch includes all of the changes to the underlying MatchMakerObject model so it will work with the Enterprise persistence layer. As the object model changed and we planned to replace Hibernate with an Enterprise server the Hibernate code was replaced and the projects can now be saved and loaded to XML. Since Hibernate was replaced setters for final fields were removed and empty constructors were removed to correct the model.

Use the hibernate-to-xml project in the MatchMaker SVN repository to export a hibernate repository to a set of xml files that are equivalent to exporting each project with the export feature. The projects can then be imported into a new MatchMaker session.

The src.org.json package is now in the library and no longer needs to be included in MatchMaker directly.